requirements.txt
Flask>=2.2.0
Flask-Login>=0.6.2
Flask-Mail>=0.9.1
Flask-Migrate>=4.0.4
Flask-SQLAlchemy>=3.0.3
SQLAlchemy>=2.0.0
python-dotenv>=1.0.0
requests>=2.28.0
google-generativeai>=0.3.0
gTTS>=2.3.2
SpeechRecognition>=3.10.0
psycopg2-binary>=2.9.6
Werkzeug>=2.2.0
gunicorn>=20.1.0
python-dateutil>=2.8.0
email-validator>=1.1.0
flask-wtf>=1.0.0
itsdangerous>=2.0.0